Control the failure cases better
By default, Ansible treats exit codes other than 0 as an error and
stops. This can be avoided by using `ignore_errors`. In this
situation, Ansible continues with the playbook, but it still
outputs red text that makes it appear that something has gone
wrong. In the task this commit alters, nothing was going wrong, so
the red output is undesirable. By using `failed_when`, the failure
behavior of Ansible can be more easily controlled.